Images, person data class, datasource , recycler wiev implemented
Assumed that we have a girl that uses the application. For this girl, I added the 6 people whom the girl wants match. I added Image, name and universiy attribute for each person. I added an  recycler_row for how our my items in the recyclerwiew will look like in the main part.
Activity details is the each persons view let'say. Person Adapter is the adapter class viweholder impleemntation and adapter impleentation is implemented here